Start of persistent path fun
Made simple script that was able to get into the persistent path and create a file with text in it and print it out. Will use something like this as save data in future projects. Also will mess around with making the change in text file data effect the scene like save data in a game would.