got it to read and load from the text file
Got it so errors don't happen if the data file is deleted while app is running. Also used Resource.Load so that I could Instiancate based off string name and spawn the prefab with the same name, this should also work in build.

Next step i want to do is have an option in-game to change the geometry and then have thst write to the data file so when you go back into the file it's change has been saved.